pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "interview-tracker"
version = "1.0.0"
description = "A desktop application for tracking job interview pipelines"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "customtkinter>=5.2.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "pendulum>=3.0.0",
    "plyer>=2.1.0",
    "schedule>=1.2.0",
    "pillow>=10.0.0",
    "google-api-python-client>=2.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=1.0.0",
    "gspread>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
interview-tracker = "interview_tracker.__main__:main"

[tool.hatch.build.targets.sdist]
include = ["src/interview_tracker"]

[tool.hatch.build.targets.wheel]
packages = ["src/interview_tracker"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""
